server/requirements.txt
Django==5.1.3
idna
packaging==24.1
virtualenv==20.26.5
PyJWT==2.6.0
django-environ==0.9.0
django-cors-headers==3.13.0
distlib==0.3.8
requests==2.32.3
sqlparse==0.5.2
platformdirs==4.3.6
filelock==3.15.4
asgiref==3.8.1
anyio
psycopg2==2.9.6